+++ pages/user/[id]/edit.jsx
@@ -11,11 +11,10 @@ import {
   Spacer,
   Stack,
   Textarea,
-  useToast,
 } from "@chakra-ui/react";
 import Head from "next/head";
 import Router, { useRouter } from "next/router";
-import { useRecoilState, useRecoilValue } from "recoil";
+import { useRecoilValue } from "recoil";
 // import { pushQuestion } from "../../firebase";
 import Header from "../../../src/components/Header";
 import styles from "../../../styles/Container.module.css";
